requirements.txt
alembic==1.0.0
aniso8601==3.0.2
atomicwrites==1.2.1
attrs==18.2.0
blessed==1.15.0
cached-property==1.5.1
cement==2.8.2
certifi==2018.8.24
chardet==3.0.4
click==6.7
colorama==0.3.9
dnspython==1.15.0
docker==3.5.1
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
eventlet==0.24.1
flake8==3.5.0
Flask==1.0.2
Flask-Boto3==0.3.2
Flask-JSGlue==0.3.1
Flask-JWT-Extended==3.13.0
Flask-Migrate==2.2.1
Flask-REST-JSONAPI==0.22.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SocketIO==3.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
greenlet==0.4.15
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.13.1
marshmallow-jsonapi==0.20.3
mccabe==0.6.1
monotonic==1.5
more-itertools==4.3.0
pathspec==0.5.5
pgpubsub==0.0.5
phonenumbers==8.9.15
pluggy==0.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
pytest==3.8.0
pytest-env==0.6.2
pytest-mock==1.10.0
python-dateutil==2.7.2
python-dotenv==0.9.1
python-editor==1.0.3
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
PyYAML==3.13
requests==2.18.4
s3transfer==0.1.13
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.2.11
sqlalchemy-views==0.2.1
termcolor==1.1.0
texttable==0.9.1
tzlocal==1.5.1
ua-parser==0.8.0
urllib3==1.22
user-agents==1.1.0
wcwidth==0.1.7
websocket-client==0.53.0
Werkzeug==0.14.1
WTForms==2.2.1
